# feat : performance metrics integration with calibre