--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -304,7 +304,9 @@ workflows:
           context: reaction-validation
           requires:
             - docker-build
-      - test-metrics
+      - test-metrics:
+          requires:
+            - deploy-to-ecs
       - snyk-security:
           context: reaction-validation
           requires: