Impact: **minor**
Type: **test**

## Issue
No standard way of measuring performance metrics on stock Starter Kit. 

## Solution
Integrate Calibre metric measurement as a step after successful deployment to ECS.

## Breaking changes
None

## Testing
1. Verify Calibre snapshots reach to Calibre's Dashboard
2. Verify Calibre runs on CI

Example Output (https://circleci.com/gh/reactioncommerce/reaction-next-starterkit/8269)

```
🇺🇸  Northern California, USA
9:17pm 27-Feb-2019

Performance Grade: 94
Progressive Web App Grade: 32
Best Practices Grade: 100
Accessibility Grade: 79
SEO Grade: 100

Timing

Time to First Byte
■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 595 ms

Response time
■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 807 ms

First Meaningful Paint
■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1.18 sec

Time to Interactive
■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 4.01 sec


Number of requests: 82
Total Page transferred: 3.32 MB

View the full report https://calibreapp.com/tests/b2c2bb9/46ace62
  
✔ Snapshot created: 4
```